+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add a contribute link into info settings tab
+
+### Changed
+
+- Task table component code improved
+- Improve message when there are not tasks to send
+- Improve design of the info settings tab
+
 ## [v0.6.2] - 2024-05-15
 
 ### Added